feat: Add max output tokens flag
This change introduces a new flag `--max-output-tokens` to limit the
number of tokens in the AI-generated commit message. This helps control
the length and cost of the generated message.

- Added `--max-output-tokens` flag
- Implemented token limit in AI options

+++ internal/ai/options.go
@@ -28,4 +28,4 @@ func WithShortMessageOnly(on bool) Option { return func(o *options) { o.ShortMes
 func WithEmoji(on bool) Option { return func(o *options) { o.EnableEmoji = on } }
 
 // WithMaxOutputTokens sets the maximum number of tokens in the output.
-func WithMaxOutputTokens(max int) Option { return func(o *options) { o.MaxOutputTokens = int64(max) } }
+func WithMaxOutputTokens(max int64) Option { return func(o *options) { o.MaxOutputTokens = max } }
